requirements.txt
CairoSVG==2.4.2
Pillow==7.0.0
